Add mg_print_mac()
